+++ frontend/components/feedback-task/container/fields/span-annotation/useSpanAnnotationTextFieldViewModel.ts
@@ -16,6 +16,10 @@ export const useSpanAnnotationTextFieldViewModel = ({
 }) => {
   const spanAnnotationSupported = ref(true);
   const answer = spanQuestion.answer as SpanQuestionAnswer;
+  const initialConfiguration = {
+    // TODO: Restore line.
+    allowOverlap: true, // spanQuestion.settings.allow_overlapping,
+  };
 
   const selectEntity = (entity: Entity) => {
     answer.options.forEach((e) => {
@@ -108,7 +112,7 @@ export const useSpanAnnotationTextFieldViewModel = ({
   };
 
   const highlighting = ref<Highlighting>(
-    new Highlighting(id, entityComponentFactory)
+    new Highlighting(id, entityComponentFactory, initialConfiguration)
   );
 
   watch(