+++ frontend/components/feedback-task/container/questions/form/rating/ratingMonoSelection.component.spec.js
@@ -4,6 +4,7 @@ import RatingMonoSelectionComponent from "./RatingMonoSelection.component";
 let wrapper = null;
 const $t = () => {};
 const options = {
+  stubs: ["BaseTooltip"],
   mocks: { $t },
   propsData: {
     options: [