+++ frontend/nuxt.config.ts
@@ -63,6 +63,8 @@ const config: NuxtConfig = {
 
     { src: "~/plugins/di" },
 
+    { src: "~/plugins/language" },
+
     { src: "~/plugins/plugins/axios.ts" },
     { src: "~/plugins/plugins/axios-cache.ts" },
     { src: "~/plugins/plugins/svg-icon.js" },
@@ -123,6 +125,11 @@ const config: NuxtConfig = {
         file: "en.js",
       },
     ],
+    detectBrowserLanguage: {
+      useCookie: true,
+      cookieKey: "i18n_redirected",
+      redirectOn: "root",
+    },
     lazy: true,
     langDir: "translation/",
     defaultLocale: "en",