+++ frontend/v1/infrastructure/storage/DatasetSettingStorage.ts
@@ -1,5 +1,5 @@
 import { useStoreFor } from "@/v1/store/create";
-import { DatasetSetting } from "~/v1/domain/entities/DatasetSetting";
+import { DatasetSetting } from "~/v1/domain/entities/dataset/DatasetSetting";
 import { IDatasetSettingStorage } from "~/v1/domain/services/IDatasetSettingStorage";
 
 const useStoreForDatasetSetting = useStoreFor<