+++ CHANGELOG.md
@@ -20,6 +20,10 @@ These are the section headers that we use:
 
 - Added `allow_overlapping` parameter for span questions. ([#4697](https://github.com/argilla-io/argilla/pull/4697))
 
+### Fixed
+
+- Improve performance for more than 250 labels ([#4702](https://github.com/argilla-io/argilla/pull/4702))
+
 ## [1.26.1](https://github.com/argilla-io/argilla/compare/v1.26.0...v1.26.1)
 
 ### Added