+++ CHANGELOG.md
@@ -18,6 +18,7 @@ These are the section headers that we use:
 
 ### Added
 
+- Added Allow overlap spans in the `FeedbackDataset` ([#4668](https://github.com/argilla-io/argilla/pull/4668))
 - Added `allow_overlapping` parameter for span questions. ([#4697](https://github.com/argilla-io/argilla/pull/4697))
 - Added overall progress bar on `Datasets` table ([#4696](https://github.com/argilla-io/argilla/pull/4696))
 - Added German language translation ([#4688](https://github.com/argilla-io/argilla/pull/4688))