+++ CHANGELOG.md
@@ -16,6 +16,8 @@ These are the section headers that we use:
 
 ## [Unreleased]()
 
+## [1.27.0](https://github.com/argilla-io/argilla/compare/v1.26.1...v1.27.0)
+
 ### Added
 
 - Added Allow overlap spans in the `FeedbackDataset` ([#4668](https://github.com/argilla-io/argilla/pull/4668))